Abstracted cocktail info into individual page components. Added react-native-swiper to handle page swiping and pagination.
