+++ package.json
@@ -16,12 +16,13 @@
     "@react-navigation/native-stack": "^6.9.17",
     "expo": "~50.0.4",
     "expo-font": "~11.10.2",
+    "expo-splash-screen": "~0.26.4",
     "expo-status-bar": "~1.11.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-native": "0.73.2",
-    "react-native-web": "~0.19.6",
-    "expo-splash-screen": "~0.26.4"
+    "react-native-swiper": "^1.6.0-rc.3",
+    "react-native-web": "~0.19.6"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
